# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shelfie API",
    "description": "图书管理 API。Metadata 写路径：PATCH 更新 DB 后异步写回 metadata.json；refresh 默认以 FS 为准。",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/libraries": {
      "get": {
        "tags": [
          "Libraries"
        ],
        "operationId": "list_libraries",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LibraryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Libraries"
        ],
        "operationId": "create_library",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLibraryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/libraries/{id}/members": {
      "get": {
        "tags": [
          "Libraries"
        ],
        "operationId": "list_members",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LibraryMemberDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Libraries"
        ],
        "operationId": "add_member",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/libraries/{id}/members/{user_id}": {
      "delete": {
        "tags": [
          "Libraries"
        ],
        "operationId": "remove_member",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/libraries/{id}/members/{user_id}/permissions": {
      "patch": {
        "tags": [
          "Libraries"
        ],
        "operationId": "update_permissions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberPermissionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/users/{user_id}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "operationId": "delete_user",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/users/{user_id}/memberships": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "list_user_memberships",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserLibraryMembershipDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/users/{user_id}/password": {
      "patch": {
        "tags": [
          "Users"
        ],
        "operationId": "change_password",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AddMemberRequest": {
        "type": "object",
        "required": [
          "role"
        ],
        "properties": {
          "can_delete": {
            "type": "boolean"
          },
          "can_edit": {
            "type": "boolean"
          },
          "can_view": {
            "type": "boolean"
          },
          "role": {
            "type": "string"
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "与 username 二选一"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ChangePasswordRequest": {
        "type": "object",
        "required": [
          "new_password"
        ],
        "properties": {
          "current_password": {
            "type": [
              "string",
              "null"
            ],
            "description": "修改自己的密码时必填"
          },
          "new_password": {
            "type": "string"
          }
        }
      },
      "CreateLibraryRequest": {
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      },
...
      "LibraryDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "root_path",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "root_path": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      },
      "LibraryMemberDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PermissionFlags"
          },
          {
            "type": "object",
            "required": [
              "user_id",
              "username",
              "role"
            ],
            "properties": {
              "role": {
                "type": "string"
              },
              "user_id": {
                "type": "string"
              },
              "username": {
                "type": "string"
              }
            }
          }
        ]
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": [
          "token",
          "user"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserInfo"
          }
        }
      },
      "PermissionFlags": {
        "type": "object",
        "required": [
          "can_view",
          "can_edit",
          "can_delete"
        ],
        "properties": {
          "can_delete": {
            "type": "boolean"
          },
          "can_edit": {
            "type": "boolean"
          },
          "can_view": {
            "type": "boolean"
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "can_delete": {
            "type": "boolean"
          },
          "can_edit": {
            "type": "boolean"
          },
          "can_view": {
            "type": "boolean"
          },
          "library_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "普通用户必填：加入的图书馆"
          },
          "library_role": {
            "type": [
              "string",
              "null"
            ],
            "description": "在图书馆内的角色：admin（馆管理员）或 member（成员）"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "UpdateMemberPermissionsRequest": {
        "type": "object",
        "required": [
          "role",
          "can_view",
          "can_edit",
          "can_delete"
        ],
        "properties": {
          "can_delete": {
            "type": "boolean"
          },
          "can_edit": {
            "type": "boolean"
          },
          "can_view": {
            "type": "boolean"
          },
          "role": {
            "type": "string",
            "description": "馆内角色：admin（馆管理员）或 member（成员）"
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "required": [
          "id",
          "username",
          "role"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "UserInfo": {
        "type": "object",
        "required": [
          "id",
          "username",
          "role"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "UserLibraryMembershipDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PermissionFlags"
          },
          {
            "type": "object",
            "required": [
              "library_id",
              "library_name",
              "role"
            ],
            "properties": {
              "library_id": {
                "type": "string"
              },
              "library_name": {
                "type": "string"
              },
              "role": {
                "type": "string"
              }
            }
          }
        ]
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}